+++ max/kernels/test/gpu/basics/test_laneid.mojo
@@ -1,5 +1,5 @@
 # ===----------------------------------------------------------------------=== #
-# Copyright (c) 2025, Modular Inc. All rights reserved.
+# Copyright (c) 2026, Modular Inc. All rights reserved.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions:
 # https://llvm.org/LICENSE.txt
@@ -54,7 +54,9 @@ fn test_grid_dim(ctx: DeviceContext) raises:
     ctx.synchronize()
 
     for i in range(buffer_size):
-        assert_equal(output_host[i] % WARP_SIZE, i % WARP_SIZE)
+        assert_equal(
+            output_host[i] % Float32(WARP_SIZE), Float32(i % WARP_SIZE)
+        )
 
     output_host.free()
