+++ max/kernels/test/gpu/linalg/test_grouped_matmul_dynamic_scaled_fp8.mojo
@@ -1,5 +1,5 @@
 # ===----------------------------------------------------------------------=== #
-# Copyright (c) 2025, Modular Inc. All rights reserved.
+# Copyright (c) 2026, Modular Inc. All rights reserved.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions:
 # https://llvm.org/LICENSE.txt
@@ -116,7 +116,7 @@ def test_grouped_matmul_dynamic_scaled_fp8_zero_edge_case[
 
     # Set up expert_ids
     for i in range(num_expert_ids):
-        expert_ids_host_ptr[i] = i % num_experts
+        expert_ids_host_ptr[i] = Int32(i % num_experts)
 
     # Create scale buffers
     comptime static_a_scales_shape = DimList(K // BLOCK_SCALE_K, Dim())
