+++ max/kernels/test/gpu/linalg/test_stream_k.mojo
@@ -1,5 +1,5 @@
 # ===----------------------------------------------------------------------=== #
-# Copyright (c) 2025, Modular Inc. All rights reserved.
+# Copyright (c) 2026, Modular Inc. All rights reserved.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions:
 # https://llvm.org/LICENSE.txt
@@ -133,7 +133,7 @@ fn mac_loop[
         sema.wait(end_iter)
         if global_r < M and global_c < N:
             C[c_offset] += accum
-    sema.release(start_iter)
+    sema.release(Int32(start_iter))
 
 
 fn first_wave_kernel[
